+++ app/javascript/users/pages/dashboard.vue
@@ -1,9 +1,9 @@
 <template>
   <v-container fill-height class="d-flex justify-center">
 
-    <!--OrdersListCards v-if="orders.length > 0" :orders="orders" />
-    <h3 v-else-if="orders.length == 0 && !loading"
-        class="d-flex justify-center subtitle">{{ $t('order.list.none') }}</h3>
+    <CourseEventsListCards v-if="course_events.length > 0" :course_events="course_events" />
+    <h3 v-else-if="course_events.length == 0 && !loading"
+        class="d-flex justify-center subtitle">{{ $t('course_event.list.none') }}</h3>
     <v-card v-else
             elevation="0"
             class="mb-12">
@@ -12,33 +12,33 @@
         size="50"
         indeterminate>
       </v-progress-circular>
-    </v-card-->
+    </v-card>
   </v-container>
 </template>
 
 <script>
   // Components
-//  import OrdersListCards from '../orders/orders_list_cards'
+  import CourseEventsListCards from '../components/course_events/course_events_list_cards'
 
   import { mapState, mapActions } from 'vuex'
 
   export default {
     created() {
-//      this.fetchOrders()
+      this.fetchCourseEvents()
     },
     components: {
-//      OrdersListCards
+      CourseEventsListCards
     },
     data() {
       return {
       }
     },
     computed: {
-//      ...mapState('order', ['orders']),
+      ...mapState('course_event', ['course_events']),
       ...mapState('layout', ['loading']),
     },
     methods: {
-//      ...mapActions('order', ['fetchOrders'])
+     ...mapActions('course_event', ['fetchCourseEvents'])
     },
   }
 </script>